issue #11184 \ref is unable to resolve references to fields in the documented struct
problem looks like to be that the mentioned fields (`date` and `day_of_the_week`) are evaluated against the outer class of the current class (so against `example::sit` instead  of `example::sit::`).
Similar problems occur with the namespace (and probably also with modules and concepts) so this also has been fixed.