XSD warnings
In CGAL we got a number of times a warning like:
```
element memberdef: Schemas validity error : Element 'memberdef': This element is not expected. Expected is ( member )
```
when running a xsd check.

The resulting XML output gave that in a `sectiondef` part there were a number of `member`s and `memberdef`s in a unsorted order and this was not supported in the `compound.xsd`.