Incorrect waring in case of Python
For the aws cli project we got the warning:
```
awscli/botocore/vendored/six.py:779: warning: documented symbol 'print_(*args ** awscli::botocore::vendored::six::kwargs' was not declared or defined.
```
note the missing `,` and at the end the missing`)`.

This is due to the fact that there was (twice) the definition `def print_(*args, **kwargs):` and here the `(*` for python was not handled properly.