Define defined in multiple, nested, files
When having a define that is defined in multiple, nested, files the second occurrence is not recorded and thus not resolved, this leads e.g. for a line like:
```
static const uint32_t k[] ALIGNED(64) = { 0 };

static const uint8_t r[] ALIGNED(64) = { 7};
```
to a warning like:
```
md5.c:5: warning: documented symbol 'static const uint8_t r' was not declared or defined.
```