issue #11271 man page verbatim changing \" to \'
6c5b9295509e introduced ManDocVisitor::filter() which corrects verbatim \\
but converts \" to \' which is inappropriate for case DocVerbatim::Verbatim
and at least one other case.

This commit adds a bool to filter() so one can tune which calls to it
do the \" conversion.

I have no idea which other calls should not do the \" conversion so for now
that conversion is turned off for all calls by commenting it out.

Fixes: 6c5b9295509e: Issue with spacing around <programlisting>
Signed-off-by: Duncan Roe <duncan_roe@optusnet.com.au>