issue #11271 man pages incorrectly convert \" to \'
6c5b9295509e introduced `ManDocVisitor::filter()` which corrects \\
but converts \" to \' which is wrong for `case DocVerbatim::Verbatim`
and the `case` of normal text.

On the assumption that there are cases where it is corect
to convert \" to \', this commit adds a bool to `filter()`
so one can tune which calls do the \" conversion.

I could not find which other call should not do the \" conversion,
so for now that conversion is turned off for all calls by commenting it out.

Fixes: 6c5b9295509e: Issue with spacing around \<programlisting\>