Don't convert CMake environment variables when using `-x_noenv` option
When having environment CMake variables in the form of `${....}` and using `doxygen -x_noenv`  we still can get warnings like:
```
warning: argument '${EIGEN_DOXY_INTERNAL}' for option INTERNAL_DOCS is not a valid boolean value
Using the default: NO!
```
whilst in this case these environment variables should not give a warning and remain in the resulting Doxyfile. These variables are normally resolved by CMake, but when creating one's own Doxyfile based on a CMake Doxyfile.in without running CMake they should not be touched and not be converted to the default value.

(Found when doing some tests with the Eigen package).