cmake: Fix flex version check for overriding the register keyword
In #11113 a check was added to override the "register" keyword for
older versions of flex. The check was added before the find_package()
call which resulted in the workaround being active for any version
of flex. This breaks the build with recent mingw-w64 which uses "register"
in its headers.

Fix by moving the version check after the find_package() call.