Handling return in `\cite` command
When having something like:
```
/**
 * First al \cite
 * mattes2001
 */
```
we get a warning about an empty / missing label and the link is generated depending on the fact whether on another place there is something like:
```
 * \cite  mattes2001
```

The split of the `\cite` and  the label is in the commentscan not allowed but allowed in the doctokenizer.
Fixed by allowing the split also in the commentscan.