Not possible to place a `\cite` in a `\par` title
When having:
```
\par cite_based \cite lorigo2001.
```
this will lead to:
```
warning: Illegal command '@cite' found as part of a title section
```
although there is no reason prohibiting it, especially as in a `\section` it is possible to have it.

(During evaluation also the `\anchor` caused some problems with a subsequent `\par` statement (due to switching to a wrong state).