issue #11415 doxywizard replaces < and > in aliases with &lt; / &gt;
We have to distinguish for writing of the settings between
- writing to the, HTML, doxywizard "console"
- writing to file / pipe to run doxygen

In the first case characters like `<` have to be escaped, in the second case this should not happen.

This is actually a regression on:
```
Commit: d3d0a1d0a5ce23e7b1d1995dc2b69ee4b493baaa [d3d0a1d]
Date: Thursday, January 2, 2025 4:00:17 PM
String representation at "show configuration" in doxywizard

In case a string in the settings contains a `<` / `>` / `&` this can give strange results as they are literally interpreted but should be escaped.
(issue is a side note to issue #11310)
```