Handling of `TIMESTAMP` in html footer and header
In ITK: https://github.com/InsightSoftwareConsortium/ITK/issues/5190  there has been a discussion about the use of `TIMESTAMP` in the html footer (and header) in respect to reproducibility (and creation of difference sets for documentation sites).
When the date is replaced by the actual value all the html files are involved. I the ITK project a html string was used and its value was replaced by means of a JavaScript function during the load of the html page.

This technique has now been implemented in doxygen as well.