Ignore non-linkable definitions when resolving links
This avoids links not being generated if a definition in a non-linkable
file has higher priority than the one in the linkable file.

Add a flag `onlyLinkable` to `getResolvedSymbolRec` which makes it skip
over unlinkable definitions. This flag is only set in `getDefsNew`. The
result of this function is ultimately used in

- DocParser::handleLinkedWord via resolveRef
- DocRef::DocRef via resolveLink
- DocLink::DocLink via resolveLink
- linkifyText via getDefs

All these users ignore unlinkable definitions.

Fixes #11560.