Input buffer overflow
When having a, very,  long string as an argument in e.g. PHP or C++ we get am error like:
```
input buffer overflow, can't enlarge buffer because scanner uses REJECT
```
which can be overcome by enlarging the lex buffers (CMake option `enlarge_lex_buffers`) but is also possible to limit the search pattern from a `+` pattern to a number of elements `{1,n}` pattern.

For PHP the string argument should, analogous to C++ not be a copy rule but a skip rule (`scanner.l`)