Handling of backticks in bibliography
When looking at the bibliography we see that there is a difference in the handling of backticks and single quotes between LaTeX and the other formats.
In LaTeX they are translated in a similar way as nowadays is done in Markdown but in the other formats they are shown literally. The bibliography items are,correctly, not run through the markdown processor as the entries are in LaTeX style.
By replacing the backticks and single quotes by the required HTML entities in the rapper script the problem is solved.