Disabling Mathjax 4 packages
With Mathjax version 4 quite a few packages are automatically loaded, though this can lead to inconstancies with, the default use of, older versions  Mathjax versions.
An example of this is the use of the `_` inside `\text` (the default Mathjax 4 behavior is the same as with regular LaTeX, see also https://github.com/doxygen/doxygen/pull/11705#discussion_r2265208555), by disabling the `textmacros` the old behavior can be obtained again (see also https://docs.mathjax.org/en/latest/upgrading/whats-new-4.0/input.html#textmacros-enabled-by-default) which is also valide for the underscore handling inside the `\text`.

By prepending a `-` (minus sign) to the package name in the `MATHJAX_EXTENSIONS` the package will be disabled.