issue #11828 Inconsistent vertical spacing of ordered list items in generated HTML
The current situation is a bit inconsistent as depending on the number of paragraphs in an item the `<p class="startli">` is added or nor.

In case one of the items consists of multiple paragraphs the `<p class="startli">` is added for all paragraphs in the items in the current item block.
When having maximal 1 paragraph for each item the  `<p class="startlisingle">` is added so the user can (through a css file) choose a different layout for both situations.