Incorrect stripping of extension of graphics files
When having files like `base_dot.ex` and `base_dot.ex1` these files are in e.g. HTML output both showing the same output as the extension is stripped for the output (so only one e.g. `base_dot.png` file is generated).
The "extension strip" function (`makeBaeName`) was incorrect as it stripped the extension regardless its name.
The function has been replaced by a new function with content based on `stripExtensionGeneral` and `stripPath`).

The problem occurred for the graphical formats `dot`, `msc` and `dia` and the output formats `html`, `latex` and `docbook`.
(Although not affected th fix has also been applied for the output format `rtf` and the graphical format `plantuml` as a matter of consistency)