Special one line CPP comment in string
When having  special one line comment (`//!` or `///`) inside a string the rule `<*>{CPPC}[!/][^\n]*` would kick in and due to its greediness an error like:
```
```
can appear when the, remainder, of the string is to large to fit in the lex buffer.

By specifying all states except those that were already rejected this problem can be overcome.