ssue #11828 Inconsistent vertical spacing of ordered list items in generated HTML
The current situation is a bit inconsistent as depending on the number of paragraphs in an item the `<p class="startli">` is added or not.

In case one of the items consists of multiple paragraphs the `<p class="startli">` is added for all paragraphs in the items in the current item block.
Now the `<p class="startli">` is always added, though this gave some problems with checked / unchecked lists, this has been overcome by moving the "check" marker to the `<ul>`, so it is now the list marker and not the item marker, hence also for each item an own list and a correction regarding the margins.